+++ src/ytextractor.nim
@@ -1,6 +1,6 @@
 #[
   Created at: 08/03/2021 19:58:57 Tuesday
-  Modified at: 08/14/2021 11:22:47 PM Saturday
+  Modified at: 08/17/2021 12:37:47 PM Tuesday
 ]#
 
 ##[
@@ -9,11 +9,13 @@
 
 import ytextractor/video; export video
 import ytextractor/channel; export channel
-import ytextractor/core/types; export types
+
 
 when isMainModule and not defined js:
   # debug purposes
   from std/times import `$`
+  import ytextractor/exports
+  import std/json
 
   # var vid = initYoutubeVideo "jjEQ-yKVPMg".videoId
   # discard vid.update()
@@ -22,8 +24,8 @@ when isMainModule and not defined js:
   # echo extractVideo("https://www.youtube.com/watch?v=u8ZP9g-RKA8")
 
   # var chan = initYoutubeChannel "https://www.youtube.com/channel/UC3aGq0eFrvrjM4F1dLUo87A".channelId
-  var chan = initYoutubeChannel "https://www./channel/UC3aGq0eFrvrjM4F1dLUo87A".channelId
+  # var chan = initYoutubeChannel "https://www./channel/UC3aGq0eFrvrjM4F1dLUo87A".channelId
+  var chan = initYoutubeChannel "https://www.youtube.com/channel/UCsBjURrPoezykLs9EqgamOA".channelId
   # var chan = initYoutubeChannel "UC3aGq0eFrvrjM4F1dLUo87A".channelId
   discard chan.update(home)
-  # discard chan.update()
-  echo chan
+  writeFile "a.json", parseJson($channelJson(chan)).pretty